doc: add mechanism for automatically generating examples from tests
Add an update-examples script to generate an examples.rst file under
doc, based on example-*.c named tests under test. This automates the
generation of examples, makes sure they build correctly (because they
are part of the tests), and the cmdtest .stdout files double as
statically generated .rst files for when hawkmoth extension isn't
available (for example when building for readthedocs).