doc: enable intersphinx to automate documentation
This commit doesn't solve any issue in and of itself, but it allows us
to move pieces of documentation where it belongs, which should resonate
with the aim of the project itself. Actually doing so will need a bigger
effort, but it will eventually make the documentation easier to
maintain.

Centralization of documentation roots for external documentation is also
good for making link maintenance easier. Unfortunately there's no
official public offering for Clang's Python bindings' documentation
available at the moment, so no link is provided for that one.

[Jani: Drop unused autodoc and napoleon for now.]