cautodoc: make source a bit more 'pythonic'
This patch replaces the string formatting '%' for the newer 'format'
method and marks members of 'CAutoDocDirective' that are not needed
outside for name mangling, possibly future proofing the current
implementation slightly.

I also removes the non standard '__author__' like variables moving the
relevant information into the package's documentation, which remains a
stub at the moment. And it removes the encoding line which is redundant
for Python 3.