doc: clean up package documentation
This makes extensive use of cross referencing, replacing many of the ad
hoc links and literals that were used before. It also makes some
opportunistic updates to the text itself for legibility and for fixing
the occasional typo although it doesn't change its meaning.

This piece of documentation is also brought up to date on the usage
instructions.

[Jani: Minor nitpicks.]