hawkmoth: deprecate the cli tool for anything other than debug
With the test usage for the Hawkmoth cli tool gone, there's no point in
maintaining it for anything other than debugging. It's still useful for
that, but allow more flexibility (e.g. change the output on a whim if
needed) by repurposing it as a debug-only tool. Focus on the Sphinx
extension being the main purpose of the project.

Remove cli part from the main documentation accordingly, and update the
cli argument parser documentation.