hawkmoth: trivial code simplifications
The parse function is currently very long and messy, which is getting in
the way of fixing all those FIXMEs as well. Anything that removes lines
of code from this function is therefore good and this is a tiny tiny
start.

There are purely cosmetic changes which have a secondary objective of
consolidating code style.

[Jani: minor indentation changes]