util: add a docstr generator
This new generator replaces all the noisy lines inside the main parser
function for the actual generation of reST documentation strings.

This patch retains the old behaviour completely, including the same
inconsistent usage of new lines and the abundance of trailing white
space. This is addressed in a following commit.