hawkmoth: nest documentation of member elements
This patch introduces obvious visual cues about the parent-member
relationship through indentation.

This also changes the output by removing the parent out of
'parent.member' declarations which was previously necessary to create
meaningful cross reference tags. This capability is now replaced by
Sphinx's automagic resolution of nested members which fortunately
doesn't change the cross referencing semantics. So the changes are
cosmetic in nature.

In spite of these being cosmetic changes, they will be intrinsic
dependencies of the recursive parsing introduced later in this series.
Namely, with further nesting levels, this is used for trivially scaling
the original parent-member scoping.