hawkmoth: recurse over structure and union members
The concept is not fully realized yet, but now there is actual
recursion. The immediate effect of this is that nested structures are
now documented as expected. However, the results may not be the expected
ones since anonymous types are not yet supported. This is addressed in a
following patch.

[Jani: rebase due to minor conflicts]