doc: don't include README in documentation
Make a clear distinction:

* The README.rst is about the project and how to install and develop
  Hawkmoth. It's mainly for the GitHub landing page.

* The Sphinx documentation under doc is about Hawkmoth configuration and
  usage.

Make references between the two, but don't try to mix them up. This will
help in further cleanups of the documentation, at the relatively small
cost of duplicating some of the introductory text.

Add a rst2html Makefile rule to be able to check the README.rst is all
right, since it's not included in the Sphinx build. Use 'make
README.html'.