Fix relative import
This was the last relative import in the project. It had been left in
for before the transition away from Python 2 so as to not break the
functionality. However, this was the wrong way to do it in Python 2 and
we've moved on to Python 3 only already anyway.

Fixes a potential override of the Hawkmoth internals by an unsuspecting
user, however unlikely it might have been.