Add way of raising error messages from the parser
This follows the current pattern used for passing the actual parsed
results except in a separate tuple. Currently it adds only diagnostics
from Clang, but the same concept can be extended to pass other warnings
or errors (e.g.) when we find weird documentation so we prod the user in
the right direction.

Original concept: Marius Vlad