test: add diagnostic tests
This patch lists no new tests, but the existing examples provide enough
coverage for Clang related warnings and errors: we don't care for
checking every possible kind of error as long as we can assert they are
generated appropriately.

The framework is quite general though and will accept tests for any new
diagnostics we add support for.