parser: fix documentation of non prototyped functions
If the arguments list of a function is empty, Clang can not rule whether
a function is variadic or not. Previously the test was called
nonetheless only to trigger an exception in Clang. This patch prevents
that situation and will instead behave as if the function had been
declared with a single 'void' parameter.