build: add initial requirements.txt
Make it easier for everyone to work on the same set of dependencies. Add
the canonical requirements.txt with sphinx==1.8.5, in alignment with the
status quo, though this is expected to change shortly.

Also add requirements-latest.txt, with just the direct dependencies of
Hawkmoth and no versions specified i.e. latest versions.