doc: rework the fallback code on failing to import hawkmoth
The sphinx ..only directive is not as clean a conditional construct as
one would expect. The contents are parsed, and, unfortunately, warnings
are issued. For example, starting with sphinx 3:

doc/examples/example-10-macro.rst:2: WARNING: Duplicate C declaration, also defined at examples:2.

In order to remove the ..only directives for "not have_hawkmoth", we add
a mock directive extension to use instead of hawkmoth on failed
import. It will simply use the sphinx include directive to include the
corresponding .rst expected test results instead of using hawkmoth.