test: remove clang warning from documentation build
Stop using an unknown type in the test:

doc/examples.rst:127: WARNING: doc/examples/example-50-struct.c:9: unknown type name 'list_data_t'

Instead of failing, this caused the type to default to int, which is
what the test expectation is instead of list_data_t! Change the source
to int as well.

The test suite does have handling for stderr output, however when
building the docs Sphinx does not have that. So avoid constructs that
emit errors from clang in examples.

Fixes #12.