util: add helper for setting up libclang on readthedocs
Turns out there is libclang installed on RTD, but the clang python
bindings are unable to find it. Add a helper to find and configure the
library path using llvm-config, happily also available.

The bindings use libclang.so for the library name, and typically it
would be an appropriately set symlink, and Config.set_library_path()
would be enough. For whatever reason RTD does not have that symlink, so
we have to set the full path using Config.set_library_name() with
libclang.so.1 as the name.

Document RTD usage while at it.