test: test transform functionality using napoleon
Add the minimal test case for transformations using napoleon.

There's a lot to be improved, though. Having the transformation
configuration duplicated in doc/conf.py and test/conf.py, as well as
importing the latter to the pure parser test, is rather ugly. It's still
better than nothing, and can be improved upon later.