cautodoc: fix directive required_arguments
Due to a typo, required_arguments was actually never used. Directive
argument parsing was always based on 0 required arguments and 1 optional
argument which was allowed to include whitespace. The optional arguments
was then split manually.

Make required_arguments 1, and add an arbitrary limit of 100 optional
arguments (there does not seem to be an "unlimited" option). Have
docutils parser split up the whitespace separated arguments for us.