parser: log doctype and name in comment metadata
We'll want to add support for filtering documentation comments based on
symbol type and name. Log the information in comment metadata for later
processing in the directive.

For the time being, the name might not be clean. See for example
_array_fixup() and _function_pointer_fixup(). We may need to provide a
name more universally suited for filtering separately, or handle the
fixups differently.