cautodoc: add a family of new symbol type specific directives
With the plumbing in place, it's almost trivial to add a bunch of new
directives that handle specific symbol types. For example:

.. c:autofunction:: filename.c function_name

Will retrieve the documentation comment for function function_name in
filename.c. Similarly for variables, typedefs, structs, unions, enums,
and macros.

FIXME: The intention is that the members of structs and unions as well
as enumeration constants for enums are included; currently they are
not. The parsed comments will probably need to be returned
hierarchically for this.