cautodoc: cache parsed documentation comments
In preparation for supporting multiple autodoc directives accessing the
same file, cache the parsed documentation comments for later use. For
the time being, this is done on a per rst document basis only. Sphinx
parallel processing makes a global cache considerably more complicated
than the few lines that are sufficient for now.