test: add support for running extension tests using the html builder
Using the html builder does not lose information as much as plain text,
providing more accurate results. However, this is considerably slower
than plain text, even if we switch to the basic html template for
speed. To alleviate this, add a separate "full" test profile, and
provide a 'make quick-test' target that skips the html tests.