docstring: fix documentation comment line prefix removal
The regex is too greedy for comments that don't start with ' * ', and it
also doesn't take lines that differ from each other into
account. Removing prefixes of different lenghts and contents is bound to
be wrong.

Actually find the longest common line prefix that contains spaces, tabs
and stars in the documentation comment, and remove it.

Notably, what this does *not* fix is mixing tabs and spaces, for
example:

/**
 * :param foo: bar
 *<tab>continuation.
 */

The failure mode will change: Previously both the leading space and tab
were removed, leading to indentation being lost on the continuation
line, now the common prefix becomes just ' *', leading to excess space
in other lines. Sphinx and reStructuredText are fussy about whitespace,
and the user should be aware of this.

This makes running the test suite about 2% slower. The impact may be
bigger on real world longer comments. This could probably be made faster
if required.

Fixes #64.