parser: handle anonymous enums, structs and unions
There are a number of ways enums, structs and unions can be anonymous in
C. Unfortunately, Clang Cursor.is_anonymous() does not return True for
all of them. However, in each case the name is the empty string and the
cursor.type.spelling conforms to patterns:

- enum (anonymous at file.c:1:1)
- struct (anonymous at file.c:1:1)
- struct T::(anonymous at file.c:1:1)
- union (anonymous at file.c:1:1)
- union T::(anonymous at file.c:1:1)

Use a regex to identify anonymous entities.

Sphinx expects @-prefixed names for anonymous entities. However, the
parser should not be aware of Sphinx. Pass an empty string as decl_name
to denote anonymous entities, and let Docstring handle what Sphinx
needs.