doc: group and deduplicate examples by sources
Currently the examples list the same source file for each example that
uses it. Group the examples by sources and deduplicate the code
listings.

This is by no means perfect, but it's still a massive improvement, and
it's a starting point for combining and grouping more of the examples
into fewer source files.