docstring: generate @names for anonymous enum/struct/union
C supports anonymous enum/struct/union, while Sphinx expects all of them
to have names. Sphinx supports naming anonymous entities via names
prefixed with @. See [1].

Consider all enum/struct/union docstrings where decl_name is passed as
the empty string '' to be anonymous, and give them names that start with
@. Sphinx will render them [anonymous] in documentation.

The names must be both unique and stable across runs (in particular to
be reproducible for testing). Lacking a better idea, use a hash across
the documentation comment and the line number.

Sphinx would support cross-referencing anonymous entities via the
made-up names, but for this to work with Hawkmoth, we'd have to come up
with a way to specify the documentation name somehow. (The workaround
is, of course, to give the entities proper names in C!)

[1] https://www.sphinx-doc.org/en/master/usage/restructuredtext/domains.html#anonymous-entities