cautodoc: improve _get_filenames method
We need this method to return a bit more context to the caller so we
know when to do automatic name resolution or not. Returning a list
instead of yielding each file name also allows us to issue an additional
warning for the CAutoDocDirective's implementation when no file is
matched.