parser: use regex in array fixup
The simple rstrip() in array fixup assumes there's a space preceding the
dimensions. This is not always the case. In particular, an array of
pointers will have the form "T *[] name".

Convert the array fixup to a regex that looks for pairs of []'s at the
end of type instead.

Fixes #78.