docstring: improve output formatting
Spurious spaces were never a technical issue, but an annoying one.
Sphinx gets rid of these spaces when compiling the rst sources anyway,
but writing new tests often required a bit of trial and error to figure
out the expected output. This patch addresses that by making the
formatting consistent.

The last patch had the side effect of fixing such oddities within
parameter lists. This patch addresses the remaining issues. Namely:

	old version         | new version
	--------------------+----------------------
	`..c:var::  fn_ptr` | `..c:var:: fn_ptr`
	`type * name`       | `type *name`

The first one was the overtly annoying one. The second one might be
debatable as a style preference, but, biased or not, it makes it more
consistent than the previous version in pointer to pointer cases too.
`type ** name` is a bit odd after all.

Incidently, together with the previous patch, this also makes the rst
look like the compiled html (and presumably all other formats) in Sphinx
5.2.3 at least. That is not particularly important though.