parser: move all formatting insto docstring
One tiny aspect of formatting remained in the parser: the function /
macro arguments. We want to leave all formatting in one place so we can
change it completely independently from the parser.

The original goal was to not change the formatting at the same time and
thus keep the unit tests unchanged. However the inconsistencies
regarding white space were too many and reproducing them would require
parser specific information to bleed into the docstring module, which
would just defeat the purpose. As such this patch ends up fixing a few
oddities in the argument list formatting.