parser: ditch intermediate parser ErrorLevel enum
Map clang diagnostic levels directly to Python logger error levels, and
get rid of the intermediate ErrorLevel enum.

The enum also implicitly acted as a verbosity level for Sphinx app. Add
that mapping separately and explicitly. We may want to revisit that
mapping later, but for now keep it functionally equivalent to what we
had before, even if it looks a bit weird.