docstring: move line splitting to get_docstring()
Start localizing line splitting in one place in docstring. There's
already some splitting and joining going on in _get_plain_comment(), so
this helps us reduce them in the future.

Add a blank line in the end if there isn't one. This unifies the number
of blank lines in the napoleon transform test case, reducing an extra
blank line.

As the whitespace (tabs to spaces) conversion is moved earlier, we also
need to adjust a few tests to account for this. We'll further adjust
this in follow-up.