test: add test coverage and reporting
We have a lot of tests, but no idea about coverage. It's actually pretty
easy to get this with pytest-cov.

Add two new make targets:

- 'make test-coverage' for normal report in the terminal

- 'make test-coverage-html' for an additional html report

The html report is actually pretty nice to navigate to find the spots
that need testing.

Eventually, with suitable "# pragma: no cover" comments and exclusion
patterns we could conceivably add --cov-fail-under 90 or even higher to
require coverage in CI.