docstring: add support for specifying the language domain
We want to potentially change the domain between C and C++, which means
different prefixes for the directives. We assume C as default
arbitrarily and because it keeps changes to a minimum for now.

Note that we avoid the term 'linkage type' in this context on purpose
since Clang cursors have a 'linkage' member with a different meaning
even though `extern "C"` is a cursor of `LINKAGE_SPEC` kind. This has
to do with the difference between what Clang calls a 'linkage kind'
(extern or not) and a 'linkage spec' (C or C++).

Note however that, for Sphynx, macros are exclusive to the C domain,
hence why those are hard-coded. This is totally fine for us, but user
may need to remember it when making references to those symbols.