docstring: add support for specifying the language domain
We want to potentially change the domain between C and C++, which means
different prefixes for the directives. We assume C as default
arbitrarily and because it keeps changes to a minimum for now.

Note however that, for Sphinx, macros are exclusive to the C domain,
hence why those are hard-coded. This is totally fine for us, but the
user may need to remember it when making references to those symbols.