test: fix C test that breaks C++ compilation
We want to use all the C syntax files as test cases for the C++ path as
they should be compatible, minus some warnings due to (e.g.) default
constructors. However, this particular `restrict` keyword seems to be a
no go in C++ even if by some troubling quirk it seems to be accepted in
C. (Would this even be defined behaviour?)

Anyway, I think this case is overkill and sounds more like a test case
for the compiler than hawkmoth. By replacing `restrict` with another
qualifier we are still testing the same code path: nothing's lost.