test: fix C test that breaks C++ compilation
We want to use as many of the C syntax files as possible as test cases
for the C++ path. They should be compatible, minus some warnings due to
(e.g.) default constructors. However, `restrict` is a C specific keyword
which gets rejected in most cases by a C++ compiler.

We are not particularly interested in testing the limits of the
compilers, least of all when the limits fall outside the boundaries of
the standard, so we split anything related to `restrict` into its own
test unit which we can later exclude from the C++ test suite.