hawkmoth: keep track of the domain of the directives
Directives already know the domain they were called from so it's natural
to propagate that information.

At this point we also make sure that we can change the domain ion the
other user of the parser: the command line tool. Once again we make 'c'
the default to avoid widespread breakage in the tests, but otherwise
arbitrarily.