parser: add cpp domain support
The domain should be inherited from the directive that the user elects
to use, hence the new parameter. However we should make sure that it
matches what the inferred domain would be if we were to take Clang's own
interpretation of the source file (as C or C++) as reference. We do that
here making sure that we abort early with a proper message if there is
any mismatch.

Note that Clang infers the language of a source file from its extension
and fails loudly when the extension is not recognized or when there is
C++ syntax in what it deems to be C code. However, we don't want to
guess which extensions are treated by Clang as C and C++, especially as
the user may override the language by specifying the `-x` argument to
Clang. Since there is nothing in the AST that directly states whether a
translation unit has been generated from C or C++ code, we look into the
AST for an indirect clue as to how Clang interpreted the file. The only
obvious way I found was to look for any of a set of macros which are
only defined for C++ code. This is not particularly beautiful, but it
works rather well.

We currently default to the C domain to avoid widespread breakage in the
tests.