parser: add mixed domain support
C++ code may include within it `extern "C" {}` blocks whose members
should be documented in the C domain. We then look for cursor kinds of
`LINKAGE_SPEC`, which show up for any such block with a cursor spelling
'C' denoting a subgraph where all symbols have C linkage and therefore
belong to the C domain. This can be observed with:

        $ clang -Xclang -ast-dump -fsyntax-only <source>
        :
        |-LinkageSpecDecl [...] C
        : |-...
        : :

Unfortunately in our own parsing we seem to always get the cursor kind
`UNEXPOSED_DECL` with spelling '' for that same declaration for some
reason. There seems to be a gap in the cursor kind mapping within the
Python's bindings, but I have not been able to get confirmation from
upstream. For now we do it the hard way by looking into the first two
tokens directly.

There are other reports of the same issue and there is an issue upstream
where I've added some additional context:

        https://github.com/llvm/llvm-project/issues/56687