parser: add mixed domain support
C++ code may include within it `extern "C" {}` blocks whose members
should be documented in the C domain. We should then look for cursor
kinds of `LINKAGE_SPEC`, which show up for any such block with a cursor
spelling 'C'. All children of such a block belong to the C domain. This
can be confirmed with:

        $ clang -Xclang -ast-dump -fsyntax-only <source>
        :
        |-LinkageSpecDecl [...] C
        : |-...
        : :

Unfortunately in our own parsing we seem to always get the cursor kind
`UNEXPOSED_DECL` with spelling '' for that same declaration. There seems
to be a gap in the cursor kind mapping within the Python's bindings, but
I have not been able to get confirmation from upstream. For now we do it
the hard way by looking into the first two tokens directly.

There are a couple of other reports I could find of the same issue, one
of them quite old. There is an issue upstream where I've added some
additional context:

        https://github.com/llvm/llvm-project/issues/56687