hawkmoth: add name appropriate configuration options
The previous configuration options were unfortunate in that they might
be more meaningful for C than C++. Instead of changing their names to
(e.g.) some convoluted 'ccppautodoc_*', we use the extension name as a
prefix.

This effectively changes the API of the extension though so we take care
to properly warn of deprecations and allowing for a transition period.