hawkmoth: add name appropriate configuration options
The configuration option names were an unfortunate selection in that
they might be more meaningful for C than C++ (expected issue in the
future) and do not properly scope the variables as part of the
'hawkmoth' plugin within `conf.py` (a present, mild issue already).

This effectively changes the API of the extension though so we take care
to properly warn of deprecations and allowing for a transition period.

We except `cautodoc_transformations` from a similar change due to plans
of deprecating the current implementation of the transformation
mechanism entirely.