hawkmoth: default to C++ on 'cpp' domain directives
Previously we added a check to make sure that the code is being parsed
as C++ when the 'cpp' domain is chosen. This check will always be
relevant in case (e.g.) a user forces Clang to parse a source file
differently.

However, it is very common in C++ projects to use .h headers, which
Clang assumes to be C code by default and warrants a special treatment
for cases like:

	.. cpp:autodoc:: foo.h

This is a very weak default which is overridden by whichever is the
latest `-x` option passed, be it from the global options or from the
directive options.

The user could also specify global Clang options, but they would affect
the 'c' domain as well, which is not ideal if the project has mixed
languages.

Note that the opposite problem also exists. The next directive would
yield an error even if `foo.hpp` actually contained strictly valid C
code only:

	.. c:autodoc:: foo.hpp

In the C case however, no one would ever be expected to use anything
other than .c and .h files. In fact, we might obfuscate what is likely a
mistake in the docs.