parser: fix anonymous type detection
Fix inconsistent handling of anonymous types, which has been kind of
broken for a long time. This is better understood through examples.

Consider these. These behave the same in both old and new versions.

	/** docstring */		| .. c:struct:: @anonymous_<hash>
	struct { ... };			|    docstring
					|
	/** docstring */		| .. c:struct:: @anonymous_<hash>
	struct { ... } v1;		|    docstring
					|
	/** docstring */		| .. c:struct:: @anonymous_<hash>
	struct { ... } v1, v2;		|    docstring

Now compare with these typedef cases (I'm using structures in these
examples, but enumerators and unions would do as well):

	/** docstring */		| .. c:type:: t1
	typedef struct type_t t1;	|    docstring
					|
	/** docstring */		| .. c:type:: t1
	typedef struct s t1, t2;	|    docstring
					|
	/** docstring */		| .. c:struct:: s
	typedef struct s { ... } t1;	|    docstring
					|
	/** docstring */		| .. c:struct:: t1
	typedef struct { ... } t1;	|    docstring
					|
	/** docstring */		| .. c:struct:: t1
	typedef struct { ... } t1, t2;	|    docstring

In the 1st and 2nd cases we document a type, in all others we document a
structure. This is clever, and almost a good thing, but it has too many
issues. Biggest one being that we document a structure with its proper
name in the 3rd case, but in the other ones we document a new type as
the structure itself. I.e. it is impossible for the documentation reader
to know that `t1` is not actually a `struct` in the same sense
`struct s` is.

Doxygen handles (used to handle?) these cases quite poorly as well, but
more importantly, whatever we do here is either going to be way too
complex or unpredictable or wrong or all of the above, se we should let
explicitness be the rule of the land instead.

For reference, the same cases now yield:

	/** docstring */		| .. c:type:: t1
	typedef struct type_t t1;	|    docstring
					|
	/** docstring */		| .. c:type:: t1
	typedef struct type_t t1, t2;	|    docstring
					|
	/** docstring */		| .. c:struct:: s
	typedef struct s { ... } t1;	|    docstring
					|
	/** docstring */		| .. c:struct:: @anonymous_<hash>
	typedef struct { ... } t1;	|    docstring
					|
	/** docstring */		| .. c:struct:: @anonymous_<hash>
	typedef struct { ... } t1, t2;	|    docstring