extensions: move cautodoc_transformations handling to an extension
Transformations via cautodoc_transformations will be deprecated in some
future release of Hawkmoth in favor of using hawkmoth-process-docstring
event.

However, we can already move cautodoc_transformations handling to an
extension that is, for now, setup by default. In the future, we can
still keep the extension, but have users explicitly set it up in conf.py
if they need it. And later on, we can remove the extension from Hawkmoth
altogether.

This also marks the beginning of built-in extensions to Hawkmoth in
hawkmoth.ext.