doc: document extending Hawkmoth
It's difficult to document events properly with rst syntax. Sphinx
documentation uses a custom 'event' object type, but the full
implementation of it is a bit much. Just define the type, and use
py:function for the event signature within the event description. It's a
bit hackish, but gets the job done.

(It's not possible to use py:function on the event directly, because the
hyphens don't work in the function name.)