parser: add support for C++ scoped enumerators
Scoped enumerators come in two forms: `enum class` and `enum struct`,
but they are exactly the same, unlike in the classes vs structs case.
Still, Sphinx has a `cpp:enum-class` and a `cpp:enum-struct` directives.

Since referencing is not broken in any way --- all enums are referenced
with `:cpp:enum:`, scoped or not ---, the only point in looking at the
tokens to distinguish between the two forms in the source code is for
Sphinx to render `class` or `struct`. That is hardly a good reason to
descend at the token level when both mean literally the same, so we use
`cpp:enum-class` for all cases of scoped enums.