update-examples: fix namespace generation
The examples can use Sphinx C domain namespaces to avoid duplicate
symbol issues when the same symbols are included more than once. The
current approach uses a hash on the testcase filename, but inadvertently
including the full path, leading to different namespaces depending on
the directory.

Instead of just feeding a relative path to the hash function, use the it
as a basis for creating a human readable namespace.