test: stop the useless input file copying
This is embarrassing. We copy the input .c file to the Sphinx srcdir for
every extension test case. However, the extension looks at hawkmoth_root
configuration for the input file root, defaulting to the Sphinx
confdir. Which, during test execution, is the test directory. We don't
use the copied files at all. So stop copying.