test: set hawkmoth_root during extension test execution
Set hawkmoth_root for extension tests to the directory holding the
testcase yaml. We'll want the filenames in the testcase yaml to be
relative to the yaml, with hopes it's the least confusing option. We
also don't want to adjust the directive during testing depending on the
paths either.

Because the hawkmoth_root depends on the testcase yaml location, we
can't set it in conf.py.

For now, these are all the same, of course, but this will change if we
arrange the tests in a directory hierarchy.