doc: fix intersphinx reference to sphinx:extensions
The object type for the "extensions" target has changed to
"confval". It's a custom object type in the Sphinx documentation, and
intersphinx is unable to reference it unless we define that locally as
well. [1]

Turn conf.py into an extension by adding the setup() function (yes,
that's indeed possible), add the "confval" object type, and fix the
extensions reference role to match that.

[1] https://github.com/sphinx-doc/sphinx/issues/5562#issuecomment-434296574