test: give more meaningful ids to test cases
Previous patches removed filename prefixes which were the only way we
could distinguish whether a test output referred to (e.g.) the `struct`
in `test/examples`, `test/c` or `test/cpp`. Filenames were never a nice
solution anyway so we instead take the directory relative to the test
folder as a namespace for the test.