test: run parser tests also on auto* directives
Add type, name and member filtering for docstring walk in order to
support running the auto* directive tests on the parser directly, and
skip fewer test cases.

We find that we do need to adjust some expected results due to
whitespace changes that the extension test via Sphinx normalizes to the
same and ignores.

We also need to add "public" to C++ struct/class members in expected
results to match the direct parser output. The extension test via Sphinx
drops them.

This does not increase test coverage overall, but improves robustness at
least wrt whitespace in the auto* directive output.