directive: add c:autotext and cpp:autotext directives
With the introduction of directives for extracting documentation for
individual symbols, documentation comments not related to any symbols
were slightly neglected. The only way to grab them is to use c:autodoc
or cpp:autodoc.

Introduce directives for extracting documentation for individual text
documentation comments, or "top level comments" as referred to in the
parser.

The obvious downside is having to add means to reference the comments by
name, which means parsing documentation comments. To keep it simple, the
name shall be from the first alphanumeric character in the comment,
inclusive, to the next :, ., or newline, non-inclusive, with no further
parsing.