extensions: overhaul javadoc/doxygen comment parsing
Full javadoc/doxygen compatibility has never been a goal for
hawkmoth. We've always promoted using pure reStructuredText and Sphinx,
because it avoids any problematic conversions. If you want all the bells
and whistles of doxygen, you should use doxygen.

However, we have to acknowledge there are a lot of codebases full of
javadoc/doxygen style documentation comments, and a lot of people who
are familiar with that style of code documentation. Requiring the use of
reStructuredText to even try hawkmoth can be quite a hurdle.

To that end, we've always had a rudimentary regex based conversion
available. But let's face the fact, is too simple, and too difficult to
maintain or extend as it is.

Try to find a middle ground with an improved parser that understands
paragraphs, inline markup, code blocks, and the like. Make it easier to
extend. Recognize all doxygen commands (more than 180 of them!), even
though we only implement a handful, making it possible to warn about the
unimplemented ones (this is for future improvement, not done yet).