venv: install hawkmoth in editable mode
Install hawkmoth in editable mode for development and testing. Use
strict mode to only use the packages described in setup.cfg [1].

This in itself does not yet detect broken packaging, because the current
directory is in python path, and the packages can be found directly in
the hawkmoth directory.

Upgrading pip and setuptools seems to be necessary on some distros when
using --system-site-packages to avoid issues.

[1] https://setuptools.pypa.io/en/latest/userguide/development_mode.html#strict-editable-installs